Moved go projects to cmd directory
