+++ cmd/product-server/README.md
