+++ cmd/product-server/products.json
